requirements.txt
streamlit
yfinance
pandas
ta
duckduckgo-search
google-generativeai
plotly
python-dotenv
watchdog
requests

# OpenBB
openbb-core
openbb-yfinance
